+++ simple_vla/model/__init__.py
@@ -1,3 +1,4 @@
 from .simple_policy import SimpleVLAPolicy
+from .state_action_processor import ActionProcessor, StateProcessor
 
-__all__ = ["SimpleVLAPolicy"]
+__all__ = ["SimpleVLAPolicy", "StateProcessor", "ActionProcessor"]